fix(seer): Make Seer repository cleanup synchronous to prevent stale dual-writes
Extract cleanup logic from Celery tasks into synchronous helpers in
autofix/utils.py and call them directly from impl.py and
organization_repository_details.py. This closes the race window where
async Seer cleanup allowed stale preferences (with disabled/hidden
repos) to be read back from Seer and dual-written into
SeerProjectRepository before the cleanup task ran.

The tasks are kept as thin wrappers (renamed with _task suffix) for
any future async callers.
