+++ src/sentry/integrations/api/endpoints/organization_repository_details.py
@@ -19,8 +19,8 @@ from sentry.deletions.models.scheduleddeletion import CellScheduledDeletion
 from sentry.models.commit import Commit
 from sentry.models.organization import Organization
 from sentry.models.repository import Repository
+from sentry.seer.autofix.utils import cleanup_seer_repository_preferences
 from sentry.tasks.repository import repository_cascade_delete_on_hide
-from sentry.tasks.seer.cleanup import cleanup_seer_repository_preferences
 
 
 class RepositorySerializer(serializers.Serializer):
@@ -100,12 +100,10 @@ class OrganizationRepositoryDetailsEndpoint(OrganizationEndpoint):
                     repository_cascade_delete_on_hide.apply_async(kwargs={"repo_id": repo.id})
 
                     if repo.external_id and repo.provider:
-                        cleanup_seer_repository_preferences.apply_async(
-                            kwargs={
-                                "organization_id": repo.organization_id,
-                                "repo_external_id": repo.external_id,
-                                "repo_provider": repo.provider,
-                            }
+                        cleanup_seer_repository_preferences(
+                            organization_id=repo.organization_id,
+                            repo_external_id=repo.external_id,
+                            repo_provider=repo.provider,
                         )
 
         return Response(serialize(repo, request.user))
