+++ tests/sentry/integrations/api/endpoints/test_organization_repository_details.py
@@ -262,7 +262,9 @@ class OrganizationRepositoryDeleteTest(APITestCase):
             organization_id=org.id, key=build_pending_deletion_key(repo)
         ).exists()
 
-    @patch("sentry.tasks.seer.cleanup.cleanup_seer_repository_preferences.apply_async")
+    @patch(
+        "sentry.integrations.api.endpoints.organization_repository_details.cleanup_seer_repository_preferences"
+    )
     def test_put_hide_repo(self, mock_cleanup_task: MagicMock) -> None:
         self.login_as(user=self.user)
 
@@ -286,14 +288,14 @@ class OrganizationRepositoryDeleteTest(APITestCase):
 
         # Verify the cleanup task was called
         mock_cleanup_task.assert_called_once_with(
-            kwargs={
-                "organization_id": org.id,
-                "repo_external_id": "uuid-external-id",
-                "repo_provider": "github",
-            }
+            organization_id=org.id,
+            repo_external_id="uuid-external-id",
+            repo_provider="github",
         )
 
-    @patch("sentry.tasks.seer.cleanup.cleanup_seer_repository_preferences.apply_async")
+    @patch(
+        "sentry.integrations.api.endpoints.organization_repository_details.cleanup_seer_repository_preferences"
+    )
     def test_put_hide_repo_with_commits(self, mock_cleanup_task: MagicMock) -> None:
         self.login_as(user=self.user)
 
@@ -315,11 +317,9 @@ class OrganizationRepositoryDeleteTest(APITestCase):
 
         # Verify the cleanup task was called
         mock_cleanup_task.assert_called_once_with(
-            kwargs={
-                "organization_id": org.id,
-                "repo_external_id": "abc123",
-                "repo_provider": "github",
-            }
+            organization_id=org.id,
+            repo_external_id="abc123",
+            repo_provider="github",
         )
 
     def test_put_rejects_integration_id(self) -> None:
@@ -347,7 +347,9 @@ class OrganizationRepositoryDeleteTest(APITestCase):
         assert repo.provider == "integrations:github"
         assert repo.integration_id is None
 
-    @patch("sentry.tasks.seer.cleanup.cleanup_seer_repository_preferences.apply_async")
+    @patch(
+        "sentry.integrations.api.endpoints.organization_repository_details.cleanup_seer_repository_preferences"
+    )
     def test_put_hide_repo_triggers_cleanup(self, mock_cleanup_task: MagicMock) -> None:
         """Test that hiding a repository triggers Seer cleanup task."""
         self.login_as(user=self.user)
@@ -371,14 +373,14 @@ class OrganizationRepositoryDeleteTest(APITestCase):
 
         # Verify the cleanup task was called with correct parameters
         mock_cleanup_task.assert_called_once_with(
-            kwargs={
-                "organization_id": org.id,
-                "repo_external_id": "github-123",
-                "repo_provider": "github",
-            }
+            organization_id=org.id,
+            repo_external_id="github-123",
+            repo_provider="github",
         )
 
-    @patch("sentry.tasks.seer.cleanup.cleanup_seer_repository_preferences.apply_async")
+    @patch(
+        "sentry.integrations.api.endpoints.organization_repository_details.cleanup_seer_repository_preferences"
+    )
     def test_put_hide_repo_no_cleanup_when_null_fields(self, mock_cleanup_task: MagicMock) -> None:
         """Test that hiding a repository with null external_id/provider does not trigger Seer cleanup."""
         self.login_as(user=self.user)
@@ -403,7 +405,9 @@ class OrganizationRepositoryDeleteTest(APITestCase):
         # Verify the cleanup task was NOT called
         mock_cleanup_task.assert_not_called()
 
-    @patch("sentry.tasks.seer.cleanup.cleanup_seer_repository_preferences.apply_async")
+    @patch(
+        "sentry.integrations.api.endpoints.organization_repository_details.cleanup_seer_repository_preferences"
+    )
     def test_put_hide_repo_no_cleanup_when_external_id_null(
         self, mock_cleanup_task: MagicMock
     ) -> None:
@@ -430,7 +434,9 @@ class OrganizationRepositoryDeleteTest(APITestCase):
         # Verify the cleanup task was NOT called
         mock_cleanup_task.assert_not_called()
 
-    @patch("sentry.tasks.seer.cleanup.cleanup_seer_repository_preferences.apply_async")
+    @patch(
+        "sentry.integrations.api.endpoints.organization_repository_details.cleanup_seer_repository_preferences"
+    )
     def test_put_hide_repo_no_cleanup_when_provider_null(
         self, mock_cleanup_task: MagicMock
     ) -> None:
